--- a/.gitignore
+++ b/.gitignore
@@ -36,6 +36,7 @@ Thumbs.db
  
 longphase 
 htslib/*
+jemalloc/*
 configure
 config.*
 autom4te.cache/