--- a/jemalloc/bin/jemalloc.sh.in
+++ b/jemalloc/bin/jemalloc.sh.in
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+prefix=@prefix@
+exec_prefix=@exec_prefix@
+libdir=@libdir@
+
+@LD_PRELOAD_VAR@=${libdir}/libjemalloc.@SOREV@
+export @LD_PRELOAD_VAR@
+exec "$@"