--- a/jemalloc/include/jemalloc/internal/private_namespace.sh
+++ b/jemalloc/include/jemalloc/internal/private_namespace.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+for symbol in `cat "$@"` ; do
+  echo "#define ${symbol} JEMALLOC_N(${symbol})"
+done