## Changes
Included `jemalloc` as a dependency in the build configuration.

## Testing
This test compares the run times with the develop branch (commit https://github.com/twolinin/longphase/commit/1cd0c68d9f664beb65e9faa3094cb20090e28907), performing modcall at ont HG002 60x. The time format is mm:ss.


| count thread (chr\*bam parsing) | Run Time Before Optimization | Maximum Memory Before Optimization (GB) | Run Time After Optimization | Maximum Memory After Optimization (GB) |
|---------------------------------|------------------------------|-----------------------------------------|-----------------------------|----------------------------------------|
| 28 (7\*4)                      | 08:02.8                      | 117.7                                   | 05:42.8                     | 101.2                                  |
| 32 (8\*4)                      | 08:17.0                      | 129.3                                   | 05:24.6                     | 103.6                                  |
| 36 (9\*4)                      | 07:32.8                      | 141.4                                   | 04:16.3                     | 109.8                                  |
| 40 (10\*4)                     | 07:33.6                      | 149.8                                   | 04:17.2                     | 117.7                                  |
| 44 (11\*4)                     | 08:08.8                      | 160.5                                   | 03:52.3                     | 124.9                                  |
| 48 (12\*4)                     | 07:47.2                      | 171.6                                   | 03:51.9                     | 135.1                                  |
| 52 (13\*4)                     | 07:45.7                      | 176.1                                   | 03:35.8                     | 139.0                                  |
| 56 (14\*4)                     | 07:30.0                      | 192.7                                   | 03:34.5                     | 143.1                                  |
| 60 (15\*4)                     | 08:30.4                      | 196.6                                   | 03:43.6                     | 149.7                                  |
| 64 (16\*4)                     | 07:58.5                      | 207.0                                   | 03:40.0                     | 155.5                                  |
| 68 (17\*4)                     | 08:16.8                      | 212.5                                   | 03:38.9                     | 158.4                                  |
| 72 (18\*4)                     | 08:52.5                      | 221.9                                   | 03:44.7                     | 163.4                                  |
| 76 (19\*4)                     | 09:16.1                      | 227.2                                   | 03:50.8                     | 166.3                                  |